Añadiendo las funcionalidad de vaca y tratamientos

+++ fronted_hackathon/hackathon_app/lib/ui/pages/web/cow_web_page.dart
@@ -111,8 +111,7 @@ class CowWebPage extends StatelessWidget {
               "Reyna",
               "4 años",
             ],
-            onLongPress: (){},
-             
+                onLongPress: () {},
                 function: () => cowInformationPage(
                     context,
                     Cattle(
@@ -121,7 +120,7 @@ class CowWebPage extends StatelessWidget {
                       weight: 10,
                       birthDate: DateTime.now(),
                       type: '',
-                      groupId: index.toDouble(),
+                      groupId: index,
                     )),
                 urlImage: '',
                 title: 'Vaca',
